requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
tensorflow>=2.16.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
supabase>=2.4.0
python-dotenv>=1.0.0
Pillow>=10.3.0
easyocr>=1.7.1
google-generativeai>=0.5.4
groq>=0.9.0
python-multipart>=0.0.9
